Refactored Project, created new page structure maintained across all Private pages
