--- src/pages/create-recipe/CreateRecipe.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-import RecipeCreationForm from "./components/RecipeCreationForm";
-import UserRecipesViewer from "../../components/UserRecipesViewer";
-
-const CreateRecipe = () => {
-  return (
-    <div className="recipe-creation-page">
-      <div className="saved-meals-container">
-        <UserRecipesViewer />
-      </div>
-      <RecipeCreationForm />
-    </div>
-  );
-};
-
-export default CreateRecipe;
