+++ src/pages/health/components/Health.jsx
@@ -1,0 +1,13 @@
+import MacroGoalForm from "./MacroGoalForm";
+import "../macroGoals.css";
+
+const Health = () => {
+  return (
+    <div className="health-page">
+      <h1 id="health-header">Health Page</h1>
+      <MacroGoalForm />
+    </div>
+  );
+};
+
+export default Health;
